remove: 회고 생성 모달 테스트 페이지 제거 및 연결 작업
